updated booking confirmation page and flow
